+++ CHANGELOG.md
@@ -20,10 +20,12 @@
         - E:/Install/Shovel/customManifest.yml # Even local manifest support. Use with caution in controlled environment!
         ```
 
+- Consider debug mode enabled only when the `debug` config option or `SCOOP_DEBUG` is valid boolean value (`$true`, `$false`, `1`, `0`, `true`, `false`)
+    - Prevent multiple evaluations of debug mode check
 - Respect `NO_JUNCTIONS` config when resolving helper utilities
     - Fallback to executable from PATH when the utility is not installed via scoop.
 
-### [0.6.5-pre1]
+### 0.6.5-pre1
 
 - **scoop-config**:
     - Consider lessmsi as default for msi extraction